+++ packages/ipc/src/ipc.ts
@@ -25,8 +25,8 @@ function ipc({ socketTimeout = 30000, path = '/tmp/app.socket-' }: IPCConfig = {
 
     function remote(handle: IPCSocket): Remote {
       return {
-        host: handle && handle._server && handle._server._pipeName || null,
-        port: handle && handle._handle && handle._handle.fd || null,
+        host: handle?._server?._pipeName || null,
+        port: handle?._handle?.fd || null,
       };
     }