+++ packages/kalm/types.d.ts
@@ -98,6 +98,10 @@ declare module 'kalm' {
         socketTimeout?: number
     }
 
+    interface QUICConfig {
+        socketTimeout?: number
+    }
+
     interface UDPConfig {
       type?: string
       localAddr?: string