+++ packages/udp/src/udp.ts
@@ -30,7 +30,7 @@ function udp({ type = 'udp4', localAddr = '0.0.0.0', reuseAddr = true, socketTim
     }
 
     function send(handle: UDPSocketHandle, payload: RawFrame | string): void {
-      if (handle && handle.socket) {
+      if (handle?.socket) {
         handle.socket.send(JSON.stringify(payload), handle.port, handle.host);
       }
     }
@@ -41,7 +41,7 @@ function udp({ type = 'udp4', localAddr = '0.0.0.0', reuseAddr = true, socketTim
     }
 
     function disconnect(handle?: UDPSocketHandle): void {
-      if (handle && handle.socket) handle.socket = null;
+      if (handle?.socket) handle.socket = null;
       emitter.emit('disconnect');
     }