+++ packages/udp/types.d.ts
@@ -31,5 +31,5 @@ declare module '@kalm/udp' {
         disconnect: (handle: SocketHandle) => void
     }
 
-    export default function udp(config?: UDPConfig): (config?: UDPConfig) => Transport;
+    export default function udp(config?: UDPConfig): (config?: UDPConfig) => KalmTransport;
 }